% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomBracket.R
\docType{data}
\name{geom_bracket}
\alias{geom_bracket}
\alias{GeomBracket}
\title{Plot brackets}
\format{
An object of class \code{GeomBracket} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
geom_bracket(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  text_size = 12,
  line_size = 1,
  linetype = "solid",
  comparisons = NULL,
  ...
)

GeomBracket
}
\arguments{
\item{text_size}{Fontsize, in pt}

\item{line_size}{Size of the bracket line}

\item{linetype}{Format of the line.}

\item{comparisons}{A list of selected comparisons to plot}
}
\description{
Base geom to plot brackets for pairwise comparisons. Takes location of the
brackets and plot the segment. Add the label. The y position of each
bracket is automatically increased to dodge each other.
}
\keyword{internal}
