% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmulticomp-package.R
\docType{package}
\name{ggmulticomp-package}
\alias{ggmulticomp}
\alias{ggmulticomp-package}
\title{ggmulticomp: Layers For Multiple Comparison}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.myuhao.com/ggmulticomp/}
}

}
\author{
\strong{Maintainer}: Yuhao Min \email{min.yuhao@mayo.edu}

}
\keyword{internal}
