% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTukey.R
\docType{data}
\name{stat_tukey}
\alias{stat_tukey}
\alias{StatTukey}
\alias{geom_tukey}
\alias{GeomTukey}
\title{Tukey HSD Test}
\format{
An object of class \code{StatTukey} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomTukey} (inherits from \code{GeomBracket}, \code{Geom}, \code{ggproto}, \code{gg}) of length 1.
}
\usage{
stat_tukey(
  mapping = NULL,
  data = NULL,
  geom = GeomBracket,
  position = "identity",
  inherit.aes = TRUE,
  show.ns = FALSE,
  trans_value = NULL,
  ...
)

StatTukey

geom_tukey(
  mapping = NULL,
  data = NULL,
  stat = "tukey",
  position = "identity",
  inherit.aes = TRUE,
  show.ns = FALSE,
  trans_value = NULL,
  text_size = 12,
  line_size = 1,
  linetype = "solid",
  ...
)

GeomTukey
}
\arguments{
\item{show.ns}{Should non-significant results be shown}

\item{trans_value}{A function that accept a vector as input and output a transformed vector. To be applied to y values before statistics test.}

\item{text_size}{Size of the significance symbols, default 12pt}

\item{line_size}{Size of the lines spanning each group, default 1pt}

\item{linetype}{Type of the line spanning each group}
}
\description{
Perform Tukey HSD Test.
Display the adjusted p-values on the figure.
}
\keyword{datasets}
