% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmulticomp-package.R
\docType{package}
\name{ggmulticomp-package}
\alias{ggmulticomp}
\alias{ggmulticomp-package}
\title{ggmulticomp: Layers For Multiple Comparison}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Yuhao Min \email{min.yuhao@mayo.edu} [contributor]
}

}
\keyword{internal}
