% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTukey.R
\docType{data}
\name{stat_tukey}
\alias{stat_tukey}
\alias{StatTukey}
\title{Tukey HSD Test}
\format{
An object of class \code{StatTukey} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
stat_tukey(
  mapping = NULL,
  data = NULL,
  geom = GeomBracket,
  position = "identity",
  inherit.aes = TRUE,
  show.ns = FALSE,
  trans_value = NULL,
  p.format = NULL,
  ...
)

StatTukey
}
\arguments{
\item{show.ns}{Should non-significant results be shown}

\item{trans_value}{A function that accept a vector as input and output a transformed vector. To be applied to y values before statistics test.}

\item{p.format}{The format of the pvalue to display. This should be a function that accept dbl as input, and return a chr to be displayed. Default to show stars}

\item{text_size}{Size of the significance symbols, default 12pt}

\item{line_size}{Size of the lines spanning each group, default 1pt}

\item{linetype}{Type of the line spanning each group}
}
\description{
Perform Tukey HSD Test.
Display the adjusted p-values on the figure.
}
\keyword{datasets}
